fix degraded pool state message
when a node entered a degraded state, exited that specific state and then re-entered it later the MCO failed to update the pool degraded reason
to the proper one. This was due to a missing annotation check in the updateNode function leading to a state where an update was not enqueued when
the daemon state reason annotation changed.
